Add WIP headless Sveltia CMS
This is an attempt to add a Content Management System with Sveltia. To access it, add the suffix "/admin" to the deployed site URL.
It works with a locally cloned repository (currently only Chromium and Edge based web browsers are supported, no Firefox and Safari) or with a GitHub account (OATH setup via CloudFlare).